requirements.txt
# Base Dependencies
Django==5.0.6
django-environ==0.11.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
# JWT Authentication
djangorestframework-simplejwt==5.3.1
# Celery and Related Extensions
celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.6.0

# Security
django-guardian==2.4.0
pyotp==2.9.0

# Audit
django-easy-audit==1.3.6

# Static Files and Media
whitenoise==6.5.0
pillow==10.3.0
qrcode==7.4.2

# File Storage
django-storages==1.14.2
django-minio-storage==0.5.8

# Utilities and Enhancements
django-filter==23.5
django-extensions==3.2.3
django-cors-headers==4.3.1
django-unfold==0.43.0
python-decouple==3.8

# Polymorphic Models
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10

# API Documentation
drf-spectacular==0.27.2

# Channels and ASGI
channels==4.1.0
daphne==4.1.2

# Deployment
gunicorn==22.0.0

# Development Tools
setuptools==70.0.0

weasyprint==62.3
django_weasyprint==2.3.0
django-browser-reload==1.13.0
django-tailwind==3.8.0

drf-extra-fields==3.7.0

requests==2.32.3

# Image Processing
numpy==2.2.1
opencv_python_headless==4.10.0.84


#OCR Detection
pytesseract==0.3.13
langdetect==1.0.9
PyMuPDF==1.25.2

# Import Export Utilities
django-import-export==4.3.5


# Excel format
openpyxl==3.1.5
